refactor: extract domain dataclasses into modules
